+++ EOLib/PacketHandlers/StatSkill/StatskillJunkHandler.cs
@@ -6,7 +6,7 @@ using EOLib.Net;
 using EOLib.Net.Handlers;
 using System.Collections.Generic;
 
-namespace EOLib.PacketHandlers.Skill
+namespace EOLib.PacketHandlers.StatSkill
 {
     /// <summary>
     /// Sent when resetting a character's stats/skills
