+++ EOLib/PacketHandlers/StatSkill/StatskillOpenHandler.cs
@@ -6,8 +6,11 @@ using EOLib.Net;
 using EOLib.Net.Handlers;
 using System.Collections.Generic;
 
-namespace EOLib.PacketHandlers.Skill
+namespace EOLib.PacketHandlers.StatSkill
 {
+    /// <summary>
+    /// Sent when opening a skillmaster dialog
+    /// </summary>
     [AutoMappedType]
     public class StatskillOpenHandler : InGameOnlyPacketHandler
     {
